Moves requests and tests to the correct namespace
