+++ eea/src/main/java/org/web3j/protocol/eea/Eea.java
@@ -16,7 +16,6 @@ import org.web3j.protocol.Web3j;
 import org.web3j.protocol.Web3jService;
 import org.web3j.protocol.core.Request;
 import org.web3j.protocol.core.methods.response.EthSendTransaction;
-import org.web3j.protocol.eea.response.PrivGetTransactionReceipt;
 
 public interface Eea extends Web3j {
     static Eea build(Web3jService web3jService) {
@@ -24,6 +23,4 @@ public interface Eea extends Web3j {
     }
 
     Request<?, EthSendTransaction> eeaSendRawTransaction(final String signedTransactionData);
-
-    Request<?, PrivGetTransactionReceipt> privGetTransactionReceipt(final String transactionHash);
 }
