+++ eea/src/main/java/org/web3j/protocol/eea/JsonRpc2_0Eea.java
@@ -18,7 +18,6 @@ import org.web3j.protocol.Web3jService;
 import org.web3j.protocol.core.JsonRpc2_0Web3j;
 import org.web3j.protocol.core.Request;
 import org.web3j.protocol.core.methods.response.EthSendTransaction;
-import org.web3j.protocol.eea.response.PrivGetTransactionReceipt;
 
 public class JsonRpc2_0Eea extends JsonRpc2_0Web3j implements Eea {
     public JsonRpc2_0Eea(Web3jService web3jService) {
@@ -33,15 +32,5 @@ public class JsonRpc2_0Eea extends JsonRpc2_0Web3j implements Eea {
                 Collections.singletonList(signedTransactionData),
                 web3jService,
                 EthSendTransaction.class);
-    }
-
-    @Override
-    public Request<?, PrivGetTransactionReceipt> privGetTransactionReceipt(
-            final String transactionHash) {
-        return new Request<>(
-                "priv_getTransactionReceipt",
-                Collections.singletonList(transactionHash),
-                web3jService,
-                PrivGetTransactionReceipt.class);
     }
 }
