+++ eea/src/test/java/org/web3j/protocol/eea/ResponseTest.java
@@ -12,17 +12,11 @@
  */
 package org.web3j.protocol.eea;
 
-import java.util.Collections;
-
 import org.junit.Test;
 
 import org.web3j.protocol.ResponseTester;
 import org.web3j.protocol.core.methods.response.EthSendTransaction;
-import org.web3j.protocol.core.methods.response.Log;
-import org.web3j.protocol.eea.response.PrivGetTransactionReceipt;
-import org.web3j.protocol.eea.response.PrivateTransactionReceipt;
 
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -42,59 +36,5 @@ public class ResponseTest extends ResponseTester {
         assertThat(
                 ethSendTransaction.getTransactionHash(),
                 is("0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331"));
-    }
-
-    @Test
-    public void testPrivGetTransactionReceipt() {
-
-        buildResponse(
-                "{\n"
-                        + "    \"id\":1,\n"
-                        + "    \"jsonrpc\":\"2.0\",\n"
-                        + "    \"result\": {\n"
-                        + "        \"contractAddress\": \"0xb60e8dd61c5d32be8058bb8eb970870f07233155\",\n"
-                        + "        \"from\":\"0x407d73d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"to\":\"0x85h43d8a49eeb85d32cf465507dd71d507100c1\",\n"
-                        + "        \"output\":\"myRlpEncodedOutputFromPrivateContract\",\n"
-                        + "        \"logs\": [{\n"
-                        + "            \"removed\": false,\n"
-                        + "            \"logIndex\": \"0x1\",\n"
-                        + "            \"transactionIndex\": \"0x0\",\n"
-                        + "            \"transactionHash\": \"0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf\",\n"
-                        + "            \"blockHash\": \"0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d\",\n"
-                        + "            \"blockNumber\":\"0x1b4\",\n"
-                        + "            \"address\": \"0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d\",\n"
-                        + "            \"data\":\"0x0000000000000000000000000000000000000000000000000000000000000000\",\n"
-                        + "            \"type\":\"mined\",\n"
-                        + "            \"topics\": [\"0x59ebeb90bc63057b6515673c3ecf9438e5058bca0f92585014eced636878c9a5\"]"
-                        + "        }]\n"
-                        + "    }\n"
-                        + "}");
-
-        PrivateTransactionReceipt transactionReceipt =
-                new PrivateTransactionReceipt(
-                        "0xb60e8dd61c5d32be8058bb8eb970870f07233155",
-                        "0x407d73d8a49eeb85d32cf465507dd71d507100c1",
-                        "0x85h43d8a49eeb85d32cf465507dd71d507100c1",
-                        "myRlpEncodedOutputFromPrivateContract",
-                        Collections.singletonList(
-                                new Log(
-                                        false,
-                                        "0x1",
-                                        "0x0",
-                                        "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf",
-                                        "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d",
-                                        "0x1b4",
-                                        "0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d",
-                                        "0x0000000000000000000000000000000000000000000000000000000000000000",
-                                        "mined",
-                                        Collections.singletonList(
-                                                "0x59ebeb90bc63057b6515673c3ecf9438e5058bca0f92585014eced636878c9a5"))));
-
-        PrivGetTransactionReceipt privGetTransactionReceipt =
-                deserialiseResponse(PrivGetTransactionReceipt.class);
-        assertThat(
-                privGetTransactionReceipt.getTransactionReceipt().get(),
-                equalTo(transactionReceipt));
     }
 }
