+++ integration-tests/src/test/java/org/web3j/protocol/pantheon/PantheonQuickstartIntegrationTest.java
@@ -26,8 +26,8 @@ import org.web3j.crypto.TransactionEncoder;
 import org.web3j.generated.HumanStandardToken;
 import org.web3j.protocol.eea.crypto.PrivateTransactionEncoder;
 import org.web3j.protocol.eea.crypto.RawPrivateTransaction;
-import org.web3j.protocol.eea.response.PrivateTransactionReceipt;
 import org.web3j.protocol.http.HttpService;
+import org.web3j.protocol.pantheon.response.privacy.PrivateTransactionReceipt;
 import org.web3j.protocol.pantheon.response.privacy.PrivateTransactionWithPrivacyGroup;
 import org.web3j.tx.LegacyPrivateTransactionManager;
 import org.web3j.tx.PantheonPrivateTransactionManager;
