+++ pantheon/src/main/java/org/web3j/protocol/pantheon/response/privacy/PrivateTransactionReceipt.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.protocol.eea.response;
+package org.web3j.protocol.pantheon.response.privacy;
 
 import java.util.List;
 import java.util.Objects;
