+++ pantheon/src/main/java/org/web3j/tx/response/PollingPrivateTransactionReceiptProcessor.java
@@ -15,9 +15,9 @@ package org.web3j.tx.response;
 import java.io.IOException;
 import java.util.Optional;
 
-import org.web3j.protocol.eea.response.PrivateTransactionReceipt;
 import org.web3j.protocol.exceptions.TransactionException;
 import org.web3j.protocol.pantheon.Pantheon;
+import org.web3j.protocol.pantheon.response.privacy.PrivateTransactionReceipt;
 
 public class PollingPrivateTransactionReceiptProcessor extends PrivateTransactionReceiptProcessor {
     private final long sleepDuration;
